City Map Formation in html format with Python
This function helps to create a map in which the cities of certain countries are given with their respective location and some basic info. Only two libraries are needed to be installed. The data file needed is uploaded here, as well as a source of download is given inside the file. Enjoy!