City Map Creator
With these python codes, it is possible to create an HTML map containing the locations and basic information of all cities of one or multiple countries or the world. Only two libraries are needed to be installed.

The data file that the codes refer to is uploaded here as well. Furthermore, the link to download the data file from the source is given in the comments inside the python file. Enjoy!